+++ themes/theme-b2b/src/components/Tag.styles.ts
@@ -3,5 +3,5 @@ import { ThemeComponent, cva } from '@marigold/system';
 export const Tag: ThemeComponent<'Tag'> = {
   tag: cva('border-border-dark flex items-center rounded-md border p-1'),
   closeButton: cva('right-1 h-4 w-4 cursor-pointer p-0 outline-none'),
-  gridCell: cva('flex items-center justify-center gap-1'),
+  listItems: cva('flex items-center gap-1'),
 };